add command to retrieve initials from snapshot
