+++ iapyx/src/cli/args/stats/snapshot.rs
@@ -1,0 +1,54 @@
+use crate::cli::args::stats::IapyxStatsCommandError;
+use crate::stats::block0::wallets::calculate_wallet_distribution_from_initials;
+use jormungandr_lib::interfaces::Initial;
+use std::path::PathBuf;
+use structopt::StructOpt;
+#[derive(StructOpt, Debug)]
+pub struct SnapshotCommand {
+    #[structopt(long = "support-lovelace")]
+    pub support_lovelace: bool,
+    #[structopt(name = "SNAPSHOT")]
+    pub snapshot: PathBuf,
+    #[structopt(long = "threshold")]
+    pub threshold: u64,
+    #[structopt(subcommand)]
+    pub command: Command,
+}
+
+#[derive(StructOpt, Debug)]
+pub enum Command {
+    Count,
+    Ada,
+}
+
+impl SnapshotCommand {
+    pub fn exec(&self) -> Result<(), IapyxStatsCommandError> {
+        let initials: Vec<Initial> = read_initials(&jortestkit::file::read_file(&self.snapshot))?;
+
+        match self.command {
+            Command::Count => calculate_wallet_distribution_from_initials(
+                initials,
+                vec![],
+                self.threshold,
+                self.support_lovelace,
+            )?
+            .print_count_per_level(),
+            Command::Ada => calculate_wallet_distribution_from_initials(
+                initials,
+                vec![],
+                self.threshold,
+                self.support_lovelace,
+            )?
+            .print_ada_per_level(),
+        };
+
+        Ok(())
+    }
+}
+
+pub fn read_initials<S: Into<String>>(snapshot: S) -> Result<Vec<Initial>, IapyxStatsCommandError> {
+    let snapshot = snapshot.into();
+    let value: serde_json::Value = serde_json::from_str(&snapshot)?;
+    let initial = serde_json::to_string(&value["initial"])?;
+    serde_json::from_str(&initial).map_err(Into::into)
+}
