+++ iapyx/src/stats/mod.rs
@@ -1,3 +1,4 @@
 pub mod archive;
 pub mod block0;
+pub mod distribution;
 pub mod live;
